Reverting to last commit by Alejandro-Ramos
This removes the image from the SmartDashboard. An image can not be
added to the SmartDashboard because there would be no place to put the
image so that every user of the code can use it (it would be dependent
on where the user clones the repository).

Don't add images to SmartDashboard